requirements/test.txt
# Test dependencies go here.
-r base.txt
django-nose==1.2
factory-boy==2.2.1
nose==1.3.0
selenium==2.37.2
yanc==0.2.4
